AI-Resume-Analyzer/requirements.txt
streamlit
pdfminer.six
sentence-transformers
scikit-learn
groq
python-dotenv
matplotlib
nltk
